+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Feature) Add `ArangoMember` CRD auto-installer
 - (Feature) Add `ArangoBackup` CRD auto-installer
 - (Feature) Add `ArangoBackupPolicy` CRD auto-installer
+- (Feature) Add `ArangoJob` CRD auto-installer
 
 ## [1.2.13](https://github.com/arangodb/kube-arangodb/tree/1.2.13) (2022-06-07)
 - (Bugfix) Fix arangosync members state inspection